Day 2
Add car evaluation model and related documentation

A new machine learning model has been developed for the evaluation of cars. The model is trained using the car evaluation dataset and is designed to predict the acceptability of a car based on its characteristics.

Several python libraries were imported for this task including matplotlib, numpy, pandas, and torch libraries. Data was preprocessed and visualizations made with matplotlib. The PyTorch library was used to create a neural network model, which was subsequently trained and evaluated on the car evaluation dataset.